:recycle: start oracle executor refactor with analysis and dice extraction
